requirements.txt
djangorestframework==3.8.2
Django==2.0.7
django-cassandra-engine==1.5.3
xlrd==1.1.0
pandas==0.23.3
matlab==0.1
pyspark==2.3.1